+++ src/main/java/com/attica/athens/domain/agora/vote/dto/response/AgoraVoteResultResponse.java
@@ -2,10 +2,7 @@ package com.attica.athens.domain.agora.vote.dto.response;
 
 public record AgoraVoteResultResponse(
         Long id,
-        Long prosCount,
-        Long consCount
+        Integer prosCount,
+        Integer consCount
 ) {
 }
-
-
-
